requirements.txt
gspread
oauth2client
notion-client
python-dotenv